+++ homeassistant/components/hassio/update.py
@@ -5,11 +5,7 @@ from __future__ import annotations
 from typing import Any
 
 from aiohasupervisor import SupervisorClient, SupervisorError
-from aiohasupervisor.models import (
-    HomeAssistantUpdateOptions,
-    OSUpdate,
-    StoreAddonUpdate,
-)
+from aiohasupervisor.models import OSUpdate
 from awesomeversion import AwesomeVersion, AwesomeVersionStrategy
 
 from homeassistant.components.update import (
@@ -40,7 +36,7 @@ from .entity import (
     HassioOSEntity,
     HassioSupervisorEntity,
 )
-from .handler import get_supervisor_client
+from .update_helper import update_addon, update_core
 
 ENTITY_DESCRIPTION = UpdateEntityDescription(
     name="Update",
@@ -308,51 +304,3 @@ async def _default_agent(client: SupervisorClient) -> str:
     mounts = await client.mounts.info()
     default_mount = mounts.default_backup_mount
     return f"hassio.{default_mount if default_mount is not None else 'local'}"
-
-
-async def update_addon(
-    hass: HomeAssistant,
-    addon: str,
-    backup: bool,
-    addon_name: str | None,
-    installed_version: str | None,
-) -> None:
-    """Update an addon.
-
-    Optionally make a backup before updating.
-    """
-    client = get_supervisor_client(hass)
-
-    if backup:
-        # pylint: disable-next=import-outside-toplevel
-        from .backup import backup_addon_before_update
-
-        await backup_addon_before_update(hass, addon, addon_name, installed_version)
-
-    try:
-        await client.store.update_addon(addon, StoreAddonUpdate(backup=False))
-    except SupervisorError as err:
-        raise HomeAssistantError(
-            f"Error updating {addon_name or addon}: {err}"
-        ) from err
-
-
-async def update_core(hass: HomeAssistant, version: str | None, backup: bool) -> None:
-    """Update core.
-
-    Optionally make a backup before updating.
-    """
-    client = get_supervisor_client(hass)
-
-    if backup:
-        # pylint: disable-next=import-outside-toplevel
-        from .backup import backup_core_before_update
-
-        await backup_core_before_update(hass)
-
-    try:
-        await client.homeassistant.update(
-            HomeAssistantUpdateOptions(version=version, backup=False)
-        )
-    except SupervisorError as err:
-        raise HomeAssistantError(f"Error updating Home Assistant Core: {err}") from err
